Corrección Nodo
